構造を少しわかりやすく
