+++ packages/frontend/src/pages/world-editor/hooks/useAvailableEntityKinds.ts
