+++ packages/frontend/src/pages/world-editor/hooks/useAvailablePlugins.ts
