+++ packages/frontend/src/pages/world-editor/index.ts
@@ -1,0 +1,1 @@
+export { WorldEditorPage } from './WorldEditorPage';
