+++ packages/frontend/src/pages/world-editor/lib/dragHelpers.ts
